% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUIComponents.R
\name{viewAngleUI}
\alias{viewAngleUI}
\title{Create the Shiny UI controls to choose the viewAngle}
\usage{
viewAngleUI(coordLabels)
}
\description{
Create the Shiny UI controls to choose the viewAngle
}

