% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUIComponents.R
\name{viewPointUI}
\alias{viewPointUI}
\title{Create the Shiny UI controls to choose the viewPoint}
\usage{
viewPointUI(coordLabels)
}
\description{
Create the Shiny UI controls to choose the viewPoint
}

